+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.16.5
+- TransferCodec, TransferBean
+  - added debug property "name" for error messages
+
+
 ## 0.16.4
 - ApiEndpoint
   - minor error handling fix